make change on the links style in the right side bar when scroll to its section
