+++ src/components/common/ComponentContent/Content/CardContentLayout/Introduction.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+import Typography from "../../../../Typography";
+
+const Introduction = () => {
+  return (
+    <>
+      <Typography variant="h1">Card</Typography>
+      <Typography variant="body1" gutterBottom>
+        Cards contain content and actions about a single subject.
+      </Typography>
+    </>
+  );
+};
+
+export default Introduction;
