--- src/components/common/TableOfContent/Card.ts
@@ -1,19 +1,0 @@
-export interface IContent {
-  id: string;
-  title: string;
-}
-const content: IContent[] = [
-  {
-    title: "Introduction",
-    id: "Introduction",
-  },
-  {
-    title: "Basic card",
-    id: "Basic_card",
-  },
-  {
-    title: "Complex Interaction",
-    id: "Complex_Interaction",
-  },
-];
-export default content;
